Added status checking during install and update
Dedede now verifies and gives a status report on the git repo after cloning the starterkit and all submodules (kirby system folder, toolkit, and panel) after installing or updating.

Because of this, Dedede now considers any problem cloning the starterkit or initalizing the submodules (except panel) to be a fatal error.

Also overhauled the way Dedede prints this information to the screen. The new presentation reflects the status report in real-time.