Fixes #1 - .DS_Store preventing install
Added a simple snippet to delete the OS X .DS_Store file in the install target that otherwise prevents an install from executing. If the directory contains other files, the is_empty check will still fail and the install will be cancelled.

There should be no side effects to this fix other than the Finder options for the target being reset.