Switched status indicators to traffic light scheme
Originally chose to colorize the success check mark ("✓") and error X ("✗") as green and red, respectively. However, upong trying this, I noticed the typical - looked out of place, and so I replaced it with the unicode black circle ("●").

After making these changes, it only made more sense to maintain a more aesthetically pleasing UX by replacing the check and x with the circle and allow the color to determine the status.